requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
typing-extensions==4.12.2
SQLAlchemy==2.0.16
openai==0.27.4
